MediaRecommendationServer/app/requirements.txt
Django>=3.0,<4.0
django-picklefield==3.1
djangorestframework==3.14.0
numpy==1.21.6
pandas==1.3.5
scikit_learn==1.0.2
xmltodict==0.12.0
django-extensions==3.2.1
psycopg2-binary==2.9.5
